# Prevent local backend file fragmentation by file preallocation.