@DRON-666 2021-02-02
Additionally I tried various Windows-specific prealocation methods instead of the generic `os.File.Truncate()`.
For backup, all methods are the same within margin of error but results for external SMR drive are very strange.

<table role="table"><tr><th align="center">Prealloc</br>method</th>
<th colspan="2" align="center">CMR SATA HDD</th>
<th colspan="2" align="center">CMR SATA HDD</th>
<th colspan="2" align="center">CMR USB HDD</th>
<th colspan="2" align="center">SMR USB HDD</th>
</tr>
<tr><td>None</td>
<td>96.93&nbsp;+0.00%</td><td>2.16</td>
<td>256.76&nbsp;-0.00%</td><td>3.19</td>
<td>173.06&nbsp;+0.00%</td><td>2.34</td>
<td>410.82&nbsp;-0.00%</td><td>2.52</td>
</tr>
<tr><td>Truncate</td>
<td>94.25&nbsp;-2.77%</td><td>1.00</td>
<td>236.74&nbsp;-7.80%</td><td>1.00</td>
<td>166.09&nbsp;-4.02%</td><td>1.00</td>
<td>557.40&nbsp;+35.68%</td><td>1.01</td>
</tr>
<tr><td>Method5</td>
<td>94.33&nbsp;-2.68%</td><td>1.00</td>
<td>234.44&nbsp;-8.69%</td><td>1.00</td>
<td>166.21&nbsp;-3.96%</td><td>1.00</td>
<td>557.66&nbsp;+35.74%</td><td>1.01</td>
</tr>
<tr><td>Method65</td>
<td>94.75&nbsp;-2.25%</td><td>1.00</td>
<td>236.49&nbsp;-7.90%</td><td>1.00</td>
<td>165.86&nbsp;-4.16%</td><td>1.00</td>
<td>558.55&nbsp;+35.96%</td><td>1.01</td>
</tr>
</table>

None - Master branch
Truncate - This PR
Method5 - `SetFileInformationByHandle(FileAllocationInfo)`
Method65 - `SetFileInformationByHandle(FileEndOfFileInfo)` + `SetFileInformationByHandle(FileAllocationInfo)`

The data in the columns is not comparable because they correspond to different computers and backup sources.
The first sub column is the average time of 10 runs, the second is the average fragmentation.