@DRON-666 2021-02-07
Thanks a lot for the review, I'll fix it tomorrow.

>As far as I know, Truncate internally calls SetFileInformationByHandle(FileEndOfFileInfo).

No, it's FileEndOfFileInfo+FileAllocationInfo (Method65):
```
benchstat Restore_Truncate.txt Restore_Method5.txt
name           old time/op  new time/op  delta
CMR_HDD1_SATA    202s ВЄ 2%    209s ВЄ 1%  +3.40%  (p=0.000 n=10+10)
CMR_HDD2_SATA    658s ВЄ 3%    714s ВЄ 2%  +8.45%  (p=0.000 n=9+9)
CMR_HDD_USB      315s ВЄ 1%    317s ВЄ 1%  +0.83%  (p=0.002 n=9+9)
SMR_HDD_USB      370s ВЄ 8%    371s ВЄ 2%    ~     (p=0.549 n=9+10)

benchstat Restore_Truncate.txt Restore_Method65.txt
name           old time/op  new time/op  delta
CMR_HDD1_SATA    202s ВЄ 2%    202s ВЄ 2%   ~     (p=0.796 n=10+10)
CMR_HDD2_SATA    658s ВЄ 3%    663s ВЄ 2%   ~     (p=0.278 n=9+10)
CMR_HDD_USB      315s ВЄ 1%    318s ВЄ 5%   ~     (p=1.000 n=9+9)
SMR_HDD_USB      370s ВЄ 8%    374s ВЄ 6%   ~     (p=0.400 n=9+10)
```
>Could you try what happens when there's only a single filestorer or when the local backend does not fsync?

Yes, of course, but every test run takes many hours, so give me a few days.