@DRON-666 2021-02-09
@MichaelEischer TL;DR By disabling fsync, you will get a speedup of two or three times. Single filestorer is a bad thing on SMR.
<table role="table"><tr><th align="center"></th>
<th colspan="2" align="center">CMR SATA HDD</th>
<th colspan="2" align="center">SMR USB HDD</th>
</tr>
<tr><td>None</td>
<td>276.49&nbsp;+0.00%</td><td>3.19</td>
<td>419.75&nbsp;+0.00%</td><td>2.57</td>
</tr>
<tr><td>Truncate</td>
<td>254.81&nbsp;-7.84%</td><td>1.00</td>
<td>561.25&nbsp;+33.71%</td><td>1.01</td>
</tr>
<tr><td>Single</td>
<td>244.71&nbsp;-11.50%</td><td>1.00</td>
<td>708.39&nbsp;+68.76%</td><td>1.01</td>
</tr>
<tr><td>NoSync</td>
<td>130.04&nbsp;-52.97%</td><td>1.00</td>
<td>170.36&nbsp;-59.41%</td><td>1.01</td>
</tr>
<tr><td>NoSync+Single</td>
<td>130.77&nbsp;-52.71%</td><td>1.00</td>
<td>165.94&nbsp;-60.47%</td><td>1.01</td>
</tr>
</table>

None - Master branch
Truncate - This PR
Single - This PR and a single filestorer
NoSync - This PR and no fsync

```
benchstat Backup_None.txt Backup_Truncate.txt
name           old time/op  new time/op  delta
CMR_HDD2_SATA    276s ± 2%    255s ± 1%   -7.73%  (p=0.000 n=9+10)
SMR_HDD_USB      421s ± 4%    563s ± 3%  +33.69%  (p=0.000 n=10+10)

benchstat Backup_Truncate.txt Backup_NoSync.txt
name           old time/op  new time/op  delta
CMR_HDD2_SATA    255s ± 1%    130s ± 1%  -48.96%  (p=0.000 n=10+10)
SMR_HDD_USB      563s ± 3%    173s ±14%  -69.25%  (p=0.000 n=10+10)

benchstat Backup_Truncate.txt Backup_Single.txt
name           old time/op  new time/op  delta
CMR_HDD2_SATA    255s ± 1%    245s ± 1%   -3.98%  (p=0.000 n=10+10)
SMR_HDD_USB      563s ± 3%    713s ± 1%  +26.67%  (p=0.000 n=10+9)

benchstat Backup_Truncate.txt Backup_NoSync_Single.txt
name           old time/op  new time/op  delta
CMR_HDD2_SATA    255s ± 1%    131s ± 1%  -48.69%  (p=0.000 n=10+10)
SMR_HDD_USB      563s ± 3%    166s ± 8%  -70.52%  (p=0.000 n=10+9)

benchstat Backup_NoSync.txt Backup_NoSync_Single.txt
name           old time/op  new time/op  delta
CMR_HDD2_SATA    130s ± 1%    131s ± 1%   ~     (p=0.165 n=10+10)
SMR_HDD_USB      173s ±14%    166s ± 8%   ~     (p=0.278 n=10+9)
```
[Raw benchstat data.](https://gist.github.com/DRON-666/e7ef587e63f9f9617e37681d8bb52887)