@DRON-666 2023-06-09
It's not the same SMR drive (now i have only two ST6000DM003), but results are very promising:
```
             │ Backup_None.txt │        Backup_Truncate.txt         │
             │     sec/op      │   sec/op    vs base                │
CMR_HDD_SATA       56.16 ± 10%   53.50 ± 8%        ~ (p=0.101 n=10)
CMR_HDD_USB        181.4 ±  4%   165.5 ± 5%   -8.77% (p=0.000 n=10)
SMR_HDD_USB        351.0 ±  1%   247.8 ± 2%  -29.40% (p=0.000 n=10)
```